twitter-alhrkn/requirements.txt
requests
requests-oauthlib
python-dotenv
streamlit
